Add reference for external repo

+++ apt/repo.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
 apt-get install -y ubuntu-cloud-keyring
-cp /root/scripts/sources.list /etc/apt/sources.list
+if [ -f /root/shared/sources.list ]
+then
+  cp /root/scripts/sources.list /etc/apt/sources.list
+fi
 apt-get update -y && apt-get dist-upgrade -y
 
 apt-get install -y crudini python-mysqldb python-openstackclient
